requirements.txt
faker==24.0.0
flask==3.0.0
click==8.1.7
pyyaml==6.0.1
qrcode==7.4.2
pillow==10.2.0
tabulate==0.9.0 